requirements.txt
absl-py==0.5.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bleach==3.0.2
blessings==1.7
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
decorator==4.3.0
dill==0.2.8.2
easydict==1.9
entrypoints==0.2.3
fastai==1.0.11
fastprogress==0.1.10
gast==0.2.0
gpustat==0.5.0
grpcio==1.12.1
h5py==2.8.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.2
nvidia-ml-py3==7.352.0
olefile==0.46
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.6.0
psutil==5.4.7
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.2.2
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
regex==2018.8.29
requests==2.19.1
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
spacy==2.0.16
tensorboard==1.11.0
tensorboardX==1.4
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==6.12.0
toolz==0.9.0
torch==1.0.0.dev20181022
torchvision==0.2.1
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
Twisted==18.9.0
typing==3.6.4
ujson==1.35
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
zope.interface==4.5.0
